requirements.txt
altgraph==0.17.2
art==5.4
future==0.18.2
MouseInfo==0.1.3
numpy==1.22.1
opencv-python==4.5.5.62
pefile==2021.9.3
Pillow==9.0.0
prettytable==3.0.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytweening==1.0.4
pywin32==303
pywin32-ctypes==0.2.0
six==1.16.0
tinyaes==1.0.2
wcwidth==0.2.5
